Add Initiate Error Catching

+++ CUTRUNTools/install_cutruntools.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
+set -e -u -o pipefail
+exec 2>&1
 START_DIR=$PWD
 SOURCE_URL="https://zenodo.org/record/3374112/files/cutruntools.tar.gz?download=1 "
 TAR_NAME="cutruntools.tar.gz"
@@ -17,7 +19,11 @@ tar -xf ${TAR_NAME}
 echo "Fetching kseq_test related files."
 cp ./cutruntools/*kseq* ./
 echo "Building kseq_test executable..."
-bash ./make_kseq_test.sh
+gcc -O2 kseq_test.c -lz -o kseq_test
+if [ ! -e "./kseq_test" ]; then
+    echo "kseq_test Installation Failure."
+    exit 1
+fi
 echo "Cleaning up source files..."
 rm -r cutruntools ${TAR_NAME}
 echo "Done."
